move indexers out of controller.go
